For multhreaded worker adding that one worker can update the access token
Summary:
We have found in the S385274 that there is some con-currency issue when multiple thread try to access or update the Storage access token (BAT/SAT). This change currently introduce a lock mechanism to address this issue.

Next steps:
We have also seen that currently we call WS token service multiple time and override the token instead we can add a way to ensure that token is only fetched once. But this might require some changes in the way we use koski. As it might add an additional step to resolve the token once.

Reviewed By: Magoja

Differential Revision: D52144115

fbshipit-source-id: e49331ab51d8544969a613f8beb7405e42da0e79