Switching to lazy stats
Summary:
With this change, we now use lazy stats instead of eager stats.
This affects both "ad-hoc" encoding selection, and "replay" encoding selection.

I measured the performance of this change for "ad-hoc" encoding selection, and it seems negligible.

Transforming a file with eager stats took 1m28s on average and the same was observed for lazy.

Performance of captured encodings transform shows a big perf improvement, at 1m08s on average (~20% improvement).

Reviewed By: Sullivan-Patrick

Differential Revision: D50943760

fbshipit-source-id: d602480d2df125724dfef9283146e650f2d00da8