Add tests for chunked stream decoder
Summary:
As part of a different diff, I noticed that ChunkedStreamDecoder (formerly known as StreamInputDecoder) doesn't have test coverage, although this class is quite complex and can be broken easily.

With this change, I added comprehensive tests to verify all the funtionality of the class.

I also used this opportunity to rename the class to ChunkedStreamDecoder to better describe what the class is doing.

Differential Revision: D53146015

fbshipit-source-id: 5520c3ceca49126b9156ce76c7a5b41917753013