Move chunked streams and add tests
Summary:
In an attempt to add tests to ChunkedStreamDecoder (formerly known as StreamInputDecoder), I noticed that ChunkedStream (new name for StreamInput), is not symmetrical (and thus, not easily testable).
By not symmetrical, I mean that creating a chunked stream happened in the TAblet, but reading a chunked stream happened somewhere else.

It doesn't actually make sense to have chunked streams in the Tablet (as tablet should not know about stream content), therefore, I moved ChunkedStream to the velox level, and created a writer class for it.

I also used this opportunity to rename StreamInput to ChunkedStream, as I think it better describes what this class is doing (e.g. handling a stream that is chunked).

Differential Revision: D53146016

fbshipit-source-id: 7cd63149a0cd9df4b451398e9008b6e31693226d