Eliminate unnecessary null allocations
Summary:
Profiling Alpha, we dicovered a performance regression because we unnecessarily allocating null buffewrs, even if the data is not nullable (or scattered), and then we proceed to release these buffers (if the data wasn't nullable), leading to a lot of waste and memory poressure.

With this change, we now push down null allocations (or reuse) into the encodings, and only use the null buffers if the encoding needs them.

Profiling the modified code shows that this bottleneck is now removed.

Differential Revision: https://www.internalfb.com/diff/D52706694?entry_point=27

fbshipit-source-id: 45e5c438828e623b2e0b30c702e388c608aea7a5