Remove chunking from FieldWriter
Summary:
Currently, stream chunking is implemented in the FieldWriter.

However, field writer shouldn't know anything about how data is layed out in the file.
This is the responsibility of the VeloxWriter layer.

This change creates a clear separation between how FieldWriter is accumulating data and how VeloxWriter is later encoding/chunking the data.

To do so, we introduce the concept of StreamData. FieldWriter is accumulating raw data in StreamData objects, and VeloxWriter is using these objects to decide when and how to flush.

As part of this change, I had to also fix chunking for string types. Previously it was easy to disable string chunking, as FieldWriter had the context to identify string sata easily. With the new implementation, it is not that easy, so I opted for fixing string chunking.

With this change I also introduced min size (configurable) for chunking, which prevents creating chunks that are too small.

Reviewed By: HuamengJiang

Differential Revision: D54879575

fbshipit-source-id: b8fa6bca229cb13b886cfa925eb5bbddecb600a8