Fix array with offsets unnecessary reallocation of buffers
Summary:
Profiling Alpha exposed an issue in the ArrayWithOffsets, where it was taking a reference to an internal vector (the internal values vector), and then verifying reference counting.
This led to the code believing that the vector is being used, so the code allocated a new vector.

With this change, we are not copying the shared ptr and thus not increasing the ref counting.

Reviewed By: HuamengJiang

Differential Revision: D52706695

fbshipit-source-id: 4e87ce2a7210385950b40c46d92b9eafb26aab7d