Change TypeWithId to take exclusive ownership on children to avoid confusion
Summary:
X-link: https://github.com/facebookincubator/velox/pull/9244

A `TypeWithId` must be a proper tree without diamonds because each node references its parent.  The constructor is reparenting the children, so taking shared ownership on children is misleading, especially when they are `const` qualified.

With this change, the constructor will take `unique_ptr` of children so they will not be shared between multiple parents.  Once the tree is constructed, the individual nodes is immutable and can be shared, so they are stored as `shared_ptr<const TypeWithId>` internally.

bypass-github-export-checks

Reviewed By: oerling

Differential Revision: D55318609

fbshipit-source-id: 12098e76110e59d364fe7eaf8448cae46103b7ed