Make reader lifecycle tests stable
Summary:
Our reader lifecycle tests verify that internal velox vectors and buffers are being reused as much as possible (as long as the caller is not holding a reference to these buffers).

In these tests we verify that the pointer before and after calling next() is the same.
However, for null buffers, they might still change, depending if the data contains nulls, vs not containing any nulls (in this case, we free the nulls vector).

Since we randomly fuzz input buffers, we sometimes get parts of the vector containing no nulls and sometimes containing nulls, leading to a nulls buffer release/allocation, and this is failing the test.

I changed the tests to only verify null pointers, if both sides are not null.

Reviewed By: sdruzkin

Differential Revision: D56361752

fbshipit-source-id: f6bc6b473c289bdd6ad88cace5c221fa6bf7886d