Generic SlidingWindowMap Reader
Summary:
### Context about SlidingWindowMap

The motivation of `SlidingWindowMap` is to introduce dictionary encoding to id_score_list feature which is represented as `map` in velox. We want to be able to deduplicate id_score_list in storage and also when reading it back.

```
MapVector of id_score_list_features

offsets         keys        vals
0                1           0.1
2                2           0.2
4                1           0.1
6                2           0.2
                 3           0.3
                 4           0.4
                 3           0.3
                 4           0.4
SlidingWindowMapWriter writes into file

offsets  lengths       keys        vals
0          2            1           0.1
0          2            2           0.2
2          2            3           0.3
2          2            4           0.4

SlidingWindowMapReader reads from file and put the deduplicated map into a Dictionary<Map>

Dictionary<Map>
                                           Map
indicies     |     offsets         lengths       keys           values
0            |       0              2             1               0.1
0            |       2              2             2               0.2
1            |                                    3               0.3
1            |                                    4               0.4
```

## This diffs

(1) Add a reader that can read back the data regardless of whether it is deduplicated or not

Reviewed By: helfman

Differential Revision: D56612236

fbshipit-source-id: 74dab19e5265bb38768f9de35e43221003d88419