Fix internal compression integration regression (#59)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/59

With the recent changes related to open sourcing Nimble, we moved internal compressors to be compiled using conditional macros.

Since this affects a header file, this macro needs to be defined on EVERY project that includes this header.

We are now switching this to not include internal compressors if a macro is defined.
It is easier to setup a global macro in the CMake system, than it is on our internal build (and less error prone).

Reviewed By: sdruzkin

Differential Revision: D58037836

fbshipit-source-id: 97cce505caa8005d92ac29621f4f54541cbee09d