Transform Optimizations + Trainer Role (#68)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/68

To better utilize our workers, I enabled the following:
1. Parallel encoding
2. Parallel stream reading
3. I/O decoupling

These increased our CPU utilization on workers from ~30% to ~85%, speeding up transforms by a lot.

In addition, created a new role to train encoding layouts for jobs (with enough items in them).
This trainer is not super robust right now, and will be improved in the future (for example, detect if it is stuck, and restert).
Not sure how much this is contributing to transform speed yet.

Reviewed By: sdruzkin

Differential Revision: D59125380

fbshipit-source-id: 6d0d2ef3bd34ba268719353238d6b8fd176e8446