Fix VeloxReader Initialization Bug (#81)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/81

The CLANGTIDY linter identified a use-after-move bug in the initialization of the `VeloxReader` executor, which was being initialized by a variable that had already been moved earlier in the initializer list. This can cause undefined behavior and in this case led to `barrier_` being given a null value, preventing the fieldReader from performing parallel reads.

After fixing this bug, some unit tests that used parallel reads started failing. Further investigation revealed that the executor in the `FieldReader` was being used incorrectly. Its executor depended on `bitmapPtr`, a pointer to a variable that went out of scope before the end of the executor contexts. This caused a stack-use-after-return exception, as described in this stack trace P1570849149.

This change addresses the incorrect initialization and the resulting exception.

Reviewed By: Yuhta

Differential Revision: D62285132

fbshipit-source-id: 4504c5c5fac97ba785fb66b18123d6c9af8278a5