Add getFileRawDataSize Test Utilities (#73)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/73

# Changes

Adding functions to the TestUtils.h in tests folders of both `encodings` and `velox` which allows us to calculate the raw data size of files by reading.
- Note the special handling for Nullable encoding types and String data types in the `getRawChunkSize` function.

# Context
This is necessary to check the correctness of the raw data size stat added in the following diff, which calculates the size when writing.
This change will also be used by nimble_dump later on (to be implemented in a later diff).

Differential Revision: D61050482

Reviewed By: sdruzkin, helfman