Add getRawDataSize to Test Utilities (#73)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/73

# Changes

Adding functions to the TestUtils.h in tests folders of both `encodings` and `velox` which allows us to calculate the raw data size of files by reading.
- Note the special handling for Nullable encoding types and String data types in the `getRawDataSize` function.

# Context
This will be used by nimble_dump later on.
RawDataSize is for streams and encodings, so this captures the raw data size *after* some optimizations such as deduping are applied, but *before* compression and serde.

Reviewed By: sdruzkin, helfman

Differential Revision: D61050482