Add callback functionality to Options/Context to capture visits to getLocalDecodedVector (#87)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/87

As title, in order to verify passthrough code is being correctly called and the local decoded vector is NOT being accessed, we need a way to track visits to the decode function. On suggestion/idea by helfman, we've created a visit callback passed in via the writer options which allows us to track if the decode visit has been called from the writer code.

I did verify that the callback is being called through some tests that I know _do_ use the decode functionality. I plan to add decode visit counters for those tests later but  that is out of scope for this diff for now

This decode visit functionality will be used in the flatmap passthrough functionality as well

Reviewed By: helfman

Differential Revision: D60241883

fbshipit-source-id: 6a415c116d11649a1f0108eed4c8ee02e2578880