Add caching support for dictionary encoded passthrough (#88)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/88

Allows for caching of the last element when a dictionary encoded array is given for passthrough.

This will bring us up to parity of local file encoding with the original arrayWithOffsets, as two consecutive datasets, either passed as unencoded array or already encoded dictionaryArray will now have the same output into the file.

Reviewed By: helfman

Differential Revision: D60413733

fbshipit-source-id: 7d78c6fb2e3f3b8394c2a8b05f7a5c48468d36f6