Remove unused-variable in /fbcode/dwio/catalog/AuditLogger.cpp (#104)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/104

[codemod] Remove unused-variable in {filename}

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient to verify
 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: palmje

Differential Revision: D65445300

fbshipit-source-id: 079b5378fc2293e5288d641d04731700e3229819