Remove unused-variable in dwio/api/test/TestApi.cpp +3 (#105)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/105

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: sdruzkin

Differential Revision: D66330367

fbshipit-source-id: e6f72b1e3006f94e8bc8d9ce1e05eb28f300432a