Remove unused-variable in dwio/nimble/velox/tests/VeloxWriterTests.cpp +3 (#106)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/106

LLVM-15 has a warning `-Wunused-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code or (b) qualifies the variable with `[[maybe_unused]]`.

#buildsonlynotests - Builds are sufficient

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: sdruzkin

Differential Revision: D66330541

fbshipit-source-id: 35d916c04e58c25c0cf8dac9983d7375972d8bf9