Capture KeepAlive instead of Executor in WriterOptions (#115)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/115

folly::Executor::KeepAlive<> is the recommended way of holding
references to Executors, as they ensure the executor is kept alive until the
KeepAlive object is destroyed. Because of this, some folly APIs can only return
KeepAlive (and not shared_ptr), such as Global pools.

These APIs cannot use WriterOption is it takes a shared_ptr

Reviewed By: xiaoxmeng, HuamengJiang

Differential Revision: D66741079

fbshipit-source-id: 715f25bd91f17abbd7006aff7d8e203deef39569