Fixing header files (#127)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/127

A new open source requirements task T211952563 was created for Nimble.

This diff completes the remaining header fixes: P1709225538.

There are 3 CMAKE files in the list that are not Nimble source files. I will add those to the exceptions list.

TODO: A`CONTRIBUTING` file is also required. I will work with the rest of the team on this.

Differential Revision: D67958287