fix: Reuse vector in LocalPartition (#122)
Summary:
X-link: https://github.com/facebookincubator/velox/pull/12002

Pull Request resolved: https://github.com/facebookincubator/nimble/pull/122

More than 10% of the CPU are spent on the destruction of local partition output when the load is high.

Also add some optimizations for serialization.  Optimization on `ByteOutputStream::appendBool` does not show significant gain in the query in example (because they are a lot small batches), but it is net gain and would be significant in large batches, so I leave it in the code.

Reviewed By: xiaoxmeng

Differential Revision: D67742489

fbshipit-source-id: 8e70dd128f31caa7909ed7c1e2b4ac1e59d7c87d