Fix Chunk Decoder for string types (#144)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/144

When a string stream is not chunked, the materialized data contains string_views pointing to the actual stream data.

Once a new chunk is leaded, the old stream data is disposed.
This means that the string_views no longer point to a valid memory.

For this reason, we have a buffering solution in the chunked decoder (only for the string case), where before we load the next chunk, we copy all the strings to a buffer managed by the chunked decoder.

Turns out this buffering logic had a bug, and it wasn't counting correctly all the strings to copy, if scattering and nulls are mixed together.

We didn't have string tests to cover this case.

As part of this change, I added string tests (which were failing very fast before the fix).

Reviewed By: sdruzkin

Differential Revision: D70021503

fbshipit-source-id: f4d0da66f70c0a0b8c4c8e5d73810de88f42a84b