Update DecodedVector memory allocation to be reused within the same context call (#153)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/153

Per suggestion from HuamengJiang, we should prevent allocating and deallocating memory for each DecodedVector construction. We will use a pool of vectors to be used, if the pool is empty, we will construct a DecodedVector. If the pool is not empty, use a DecodedVector from the pool

The vector will be returned to the pool whenever LocalDecodedVector goes out of scope.

DecodedVectorManager will handle the lifecycle of LocalDecodedVector.

RawSizeContext is created for separation of concerns and for the chance we need to add more context independently.

## Benchmark output
NewContext is benchmarking the allocation a new DecodedVector per context.
SameContext is benchmarking the reuse of DecodedVector per context.
Results prove that reusing the same context will improve performance
```
buck run @//mode/opt //dwio/nimble/velox/tests:raw_size_benchmark
```
```
============================================================================
[...]mble/velox/tests/RawSizeBenchmark.cpp     relative  time/iter   iters/s
============================================================================
NewContext                                                 43.69ns    22.89M
SameContext                                                 9.03ns   110.68M
```

Reviewed By: HuamengJiang

Differential Revision: D71333369

fbshipit-source-id: f61d5f937f05019e920c003f45421c76711253f7