Allow nimble writer to work with leaf pool (#162)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/162

Nimble writer needs root pool because it will create agg pools to do memory accounting. However, this makes it harder for caller as most of the cases, caller doesn't know if a writer needs a leaf pool or not, hence have to know the implementation details to pass the right pool.

To make it worse, in Nimble serializer uses a root pool as well due to using Nimble writer. It's quite hard to wire up an agg pool there, and causes a lot of confusion different pools have to be wired up due to the encoding type and if it's encoding or decoding.

This diff addresses the problem by allowing Nimble to use a leaf pool. In the case when leaf pool is passed in, it will use it only for allocation without creating additional child pools.

Reviewed By: HuamengJiang

Differential Revision: D74020754

fbshipit-source-id: ba2909c3f4226085a02fa72250c469f78cd4cd4a