Fix incorrect nulls in multi-chunk sparse read (part 2) (#166)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/166

The previous fix (D73993832) only addresses the issues when the chunk
boundary does not end on a byte.  There is one more case that when the chunk
does not begin on a byte boundary, we decode the result on the next byte
boundary, then copy the result over to its final position.  If there is trailing
nulls, it could overwrite the bits of next chunk, we need to set them back after
the copy.

Reviewed By: HuamengJiang

Differential Revision: D74327433

fbshipit-source-id: a8f43bcca2ae7df30d355ee77995ac7c46647048