fix: Avoid decompress data when estimate row size
Summary:
With some very low selectivity queries (more stripes than output rows),
Nimble still needs to load and decompress the data in order to do the row size
estimation.  The shall be addressed once we have column statistics metadata; in
the meantime we can avoid the decompression cost by just get the row count and decompressed
size instead of actual decompressing the data.  This brings us on par with same
query on DWRF.

Differential Revision: D74884140