Optimize FBW writer (#173)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/173

FBW writers always allocates +7 extra bytes in the output buffer to make sure the vectorized decoder does not fail on non-padded data. This results in a fixed potentially unnecessary overhead for the storage size.

This change reduces extra buffer from 7 bytes to only required number of bytes to pad by a machine word with size of 8 bytes.

Differential Revision: D75581018