Back out "Free up memory in DictionaryEncoding::encode faster"
Summary:
The diff above will create a big memory bandwidth regression.
Nimble Buffers are not meant to be short lived.
Each buffer allocates a very big chunk of memory inside.


The proper fix (which was planned to be implemented as part of the compression fallback change) was to have two buffers to be used at the time of encodings.
One for temp memory, like this (which will be dicarded after all encodings are done), and one for holding the encoding results.

Reverting this change for now.

Differential Revision: D75824819