Extract Null Count from Raw Size Utils (#177)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/177

Nimble file-level logical size is currently calculated using methods in the RawSizeUtils file. We can extend this util to also calculate column level null counts. This diff introduces support for this by calculating the null counts of the children in the top-level row of the Velox vector. We include unit testing for this change.

Differential Revision: D75928862

Reviewed By: nathanphan26