Extract Column Logical Size from Raw Size Utils (#176)
Summary:

Nimble file-level logical size is currently calculated using methods in the RawSizeUtils file. We can extend this util to also calculate column level logical size. This diff introduces support for this by calculating the logical size of the children in the top-level row of the Velox vector. We include unit testing for this change.

Reviewed By: nathanphan26

Differential Revision: D75927903