Calculate Column Stats from Field Writers (#188)
Summary:

In the diff, we introduce a column stats class that is used to accumulate the logical size, deduped logical size, physical size, null count, and value count from the nimble FieldWriter class. We also include unit tests for these. The next step is using this information to update the in memory stats representation defined [here](https://www.internalfb.com/diff/D76456608). In the next diff we will also track the expected logical the size of keys for struct injected maps and reuse the same stats object for all flatmap keys to optimize for memory.

Reviewed By: HuamengJiang

Differential Revision: D76633052