fix: Read lazy ROW value in flat map vector returning incorrect result (#241)
Summary:
X-link: https://github.com/facebookincubator/velox/pull/14703

Pull Request resolved: https://github.com/facebookincubator/nimble/pull/241

When the following conditions are satisfied:
1. The flat map column has ROW as the map value (there is no prod data with this type; we discovered it using table evolution fuzzer)
2. The flat map column is read as ordinary map
3. The flat map column has a push down filter in the query
4. A second filter is evaluated and reducing the number of rows after flat map column

The children in the ROW map values would be read lazily; however the output row
set of the ROW column is not set correctly, because we consider its parent a MAP
column and its value should not be read lazily.  Fix this by not allowing the
MAP children to be read lazily in the flat map as map column readers.
bypass-github-export-checks

Reviewed By: juwentus1234

Differential Revision: D81390639

fbshipit-source-id: cf41fa53c2db5097e9d35d14a90fa6fc8e4f70ff