Removing mutableXXX calls from reader
Summary:
Following up on the discoveries that mutableXXX methods are sometimes inefficient (as they unnecessarily might call reallocate, which unnecessarily copies data), I decided to remove most of them from the code.

Adding to that, we don't need most of these calls, as we pre-allocate the correct size of the buffers ahead of time (during VectorInitializer), so we have to correct size already.

I also figured out that padding is not needed for nulls, as allocate alreay adds padding.

This change is a bit risky, so I verified it in validation service for few days. I also used asan-ubsan modes in validation service.

Reviewed By: sdruzkin

Differential Revision: D81613207