Fix buffer resize regression (#245)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/245

Recent diff broke the Nimble reader buffer size reallocation.

This diff fixes the issue.

I also discovered that previous and existing allocations are very wasteful, so I used this opprtunity to fix them.
Previously, the code was calling reallocate in many cases, assuming that reallocate is a no-op if there is enough capacity.

Turns out that reallocate was checking for shared ptr ref count, and ended up reallocating every time, as our code is holding two references (non-uinque).

Another issue was that  reallocate was unnecessarily copying data from old buffer to new buffer.

With the new code, we are now only using allocate (never reallocate). In the case were a resize is needed, the existing buffer is dropped, and a new one is allocated, without data copy.

Reviewed By: sdruzkin

Differential Revision: D81613204

fbshipit-source-id: 3b976f27cd515355b327346847ae80ea71015532