Free up memory held by previous stripe in batch reader (#244)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/244

Batch reader was holding up to the previous stripe streams during a loading of the next stripe. This resulted in significant increase (50-400MB) in peak memory consumption when the reader was loading the next stripe.

This change resets decoders from the previous stripe before loading the next stripe. It brings the batch reader almost on par with the reported peak memory with the selective reader.

Reviewed By: macvincent, helfman

Differential Revision: D81859545

fbshipit-source-id: 75654ee7bfc629d9544c9bcf7808112f0649ac0e