Disable MC in the compression test
Summary:
As MacVincent discovered, the MC is getting enabled in opt mode, which changes the compression behavior. With enabled MC the minimal compression threshold for MetaInternal grows by additional 20 bytes.

For now let's just disable the MC in the test to force it to the old behavior, and let's re-evaluate the MC threshold separately to see whether we should raise it to 80-100 bytes as previously discussed, or adjust to the new MC threshold.

Reviewed By: macvincent

Differential Revision: D83215102

fbshipit-source-id: f9046809a3075aba5524ce9af7019dff07a1d48f