turn CompressionParameters into a struct (#270)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/270

Because `CompressionParameters` is a union, the default initialization is to only initialize the zstd params portion. The result is that the `MetaInternalCompressionParameters` is not initialized properly in places where defaults are expected. This is the cause of the bug encountered in D83197927 and doubtless other small bugs here and there.

`CompressionParameters` was declared as a union in D40728763. I saw no reason given for this to be a union and not a struct. On the contrary, there are many use cases within the codebase where the union is treated like a struct, including in the test itself!
https://www.internalfb.com/code/fbsource/[23ec3ab74d6b6bd8bd228b1cdc9537b65e1dbc96]/fbcode/dwio/nimble/encodings/tests/CompressionTests.cpp?lines=46%2C49

Reviewed By: sdruzkin, macvincent

Differential Revision: D83626048

fbshipit-source-id: 2452d56d8f1b258ab3635c839856bfb4d9eee335