fix: Handle null constant result in duplicated array and map column readers
Summary:
In case of certain layout of deduplicated arrays or maps, specifically:

1. The deduplicated field is stored as flat map values, with at least 2 keys;
2. Array elements in one of the map values are all nulls, and thus stored as null streams;
3. The reader reads the key with null elements runs first, and the read starts a new run;
4. Then read the other key, and it is a continuation of previous run, plus some new elements at end.

When the above conditions are all satisfied, the elements vector is changed to a constant vector by `NullColumnReader` first, then we try to write to it because the second key is reusing the value from previous run, so it tries to copy the cached value into result vector, and throw an exception.  Fix this by resetting the result if it is not writable (constant or dictionary).

Reviewed By: HuamengJiang

Differential Revision: D84085512

fbshipit-source-id: 12f53414922dd2907b4a7d709ed35a08c2829683