Support Per Stream Chunking to Relieve Memory Pressure (#243)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/243

This is an implementation of a detail in the new chunking policy described in this [doc](https://fburl.com/gdoc/gkdwwju1). Rather than chunking all eligible streams, we chunk individual streams in the order of their raw size until memory pressure is relieved. For our unit tests, the maximum number of chunks produced is identical to the previous implementation. But there may be differences for large file sizes. This requires more experimentation and tuning to determine the right threshold value that takes advantage of this.

Reviewed By: helfman

Differential Revision: D81715655

fbshipit-source-id: bd989b7f619651a160ad1574419be94dceed30e4