Introduce a flag that allows ignoring top level nulls in the writer (#296)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/296

Having top level nulls is not a valid use case in the context of data warehouse. Nimble writer accepts top level nulls and write them as is, which means downstream readers that tries to read from the file may encounter exceptions or produce wrong result. In contrast, DWRF writer ignores nulls and there still dependency that may produce vectors with top level nulls. Introduce flag that controls if nulls should be ignored for the root writer, so we can ensure a seamless migration.

Reviewed By: sdruzkin

Differential Revision: D86210058

fbshipit-source-id: 3bf0c5e464aedc84957a221ac630af92786e7c23