Determine Max Chunk Size Based on Schema Width (#304)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/304

The size of a single element in the Nimble Velox writer may exceed `maxStreamChunkRawSize`. We did not handle such cases correctly for strings, leading to some failures during Vader testing. This diff fixes  a logical bug in the stream chunkers that affected how we handled such cases.  We also introduce `wideSchemaMaxStreamChunkRawSize` and `largeSchemaThreshold` writer options. This ensures we can increase the max chunk size threshold in cases where we do not expect a large number of streams, prevent chunks with a small number of elements when individual values are large, and take advantage of string dictionary encoding for such cases.

Reviewed By: sdruzkin

Differential Revision: D86435350

fbshipit-source-id: e5dcb741bbacc266dc56fa5a9052d3dabb82eaf6