fix: [Nimble][Chunking] Determine Threshold for Chunking Stages Based on Schema Width (#316)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/316

There are two chunking stages. A soft stage where we only chunk streams above a pre-determined max size and a hard stage where we chunk all streams above min size. In D86435350, we introduced an option to determine max size based on schema width. While we use this new max size value to initialize the chunker, we fail to use it to when determining chunking stages. This diff fixes that.

Reviewed By: helfman

Differential Revision: D86924613

fbshipit-source-id: 4c07eefe22d4474aa1c07cfa5227fb92d71a89b4