Reserve only Required Capacity During Compaction (#317)
Summary:

When implementing the stream chunker, we anticipated that stream buffers after chunking will end up growing to the size that previously triggered chunking. As a tradeoff between minimizing reallocations (for performance) and actually releasing memory (to relieve memory pressure), we heuristically decided to set the new buffer size to be maxChunkSize_ plus 50% of the difference between the current buffer capacity and maxChunkSize_. The problem with this is that when a majority of streams before chunking are less than maxChunkSize_, we never get to relieve memory pressure. We will be reverting this decision for now to see what the actual performance implications would be in DISCO experiments. Local tests indicate significant memory improvements for certain files.

Differential Revision: D87022554