Relieve Memory When Chunking in Hard Mode (#317)
Summary:
Pull Request resolved: https://github.com/facebookincubator/nimble/pull/317

When implementing the stream chunker, we anticipated that stream buffers after chunking will end up growing to the size that previously triggered chunking. As a tradeoff between minimizing reallocations (for performance) and actually releasing memory (to relieve memory pressure), we heuristically decided to set the new buffer size to be maxChunkSize_ plus 50% of the difference between the current buffer capacity and maxChunkSize_. However, in cases where the current capacity is less than maxChunkSize_, we retained the current capacity. This later decision is the reason we don't see memory savings in a good number of training tables.

This is because chunking occurs in two stages. A soft stage where we only chunk streams above a pre-determined max size and a hard stage where we chunk all streams above min size. The problem with our previous approach was that when a majority of streams being chunked are less than maxChunkSize_, we never get to relieve memory pressure from the memory pool. We fix that by setting the new capacity for small streams to be the max between the required capacity and 50% of the previous capacity. Since these streams are small anyway, we don't have to worry too much about the cost of memory reallocation.

With this fix, local instrumentation testing indicates that we should expect over 2X improvements in average memory relative to the default Nimble baseline. We will run DISCO experiments once this land to get more data.

Reviewed By: helfman

Differential Revision: D87022554

fbshipit-source-id: 75cbb99da58e08575db3777f051caff57ceeaa30