Introduce JK to Disable Memory Reallocation Optimization (#325)
Summary:

When implementing the stream chunker, we anticipated that stream buffers after chunking will end up growing to the size that previously triggered chunking. As a tradeoff between minimizing reallocations (for performance) and actually releasing memory (to relieve memory pressure), we heuristically determine the new buffer capacity for each stream to be larger that required. The issue with this optimization is that it conflicts with the rest of our memory tracking logic since we now have retained memory in the memory pool that is not accounted for. We now know through local testing that disabling this optimization leads to better memory pressure relief. 

We performed local DISCO tests with and without this optimization for two tables and included the comparison in https://docs.google.com/spreadsheets/d/1kZvBwhVHZRyB7tg-qT2Et4V_7mDWyNr-VrtDzjG_FKY/view?gid=1209014630#gid=1209014630. It shows that for these two tables we save an average of **15%** improvement in average write memory. We also see **4%** write CPU improvement on average. Though we saw **-6%** regression in CPU for a single 256MB stripe experiment. These results indicate that more testing is required in order to understand its impact on a larger sample size.


In this diff, we introduce a JK, [dwio/nimble_chunking:disable_memory_reallocation_optimization](https://www.internalfb.com/intern/justknobs/?name=dwio%2Fnimble_chunking#disable_memory_reallocation_optimization), that will be enabled just for DISCO experiments. This will help us understand the full impact of this optimization and whether it should be retained.

Differential Revision: D87494427